+++ packages/independent/custom-elements-redefine/scripts/dev.mjs
@@ -1,7 +1,9 @@
 import { startDevServer } from "@jsenv/core";
+import { jsenvPluginExplorer } from "@jsenv/plugin-explorer";
 
 export const devServer = await startDevServer({
   sourceDirectoryUrl: new URL("../src/", import.meta.url),
   port: 3457,
   clientAutoreload: false,
+  plugins: [jsenvPluginExplorer()],
 });
