+++ packages/independent/custom-elements-redefine/src/tests/adopted_callback.test.html
@@ -45,7 +45,7 @@
       await createAnimationFramePromise();
       assert({
         actual: adoptedCallbackValue,
-        expected: `adopted from http://localhost/ to about:blank`,
+        expected: `adopted from ${window.location.href} to about:blank`,
       });
     </script>
   </body>
