+++ packages/independent/custom-elements-redefine/src/tests/connected_callback.test.html
@@ -1,7 +1,7 @@
 <!doctype html>
 <html>
   <head>
-    <title>adoptedCallback tests</title>
+    <title>connectedCallback tests</title>
     <meta charset="utf-8" />
     <link rel="icon" href="data:," />
   </head>
