+++ packages/independent/custom-elements-redefine/src/tests/constructor.test.html
@@ -1,7 +1,7 @@
 <!doctype html>
 <html>
   <head>
-    <title>adoptedCallback tests</title>
+    <title>constructor tests</title>
     <meta charset="utf-8" />
     <link rel="icon" href="data:," />
   </head>
